chore: refactor test configs to align with `game.save` package structure
Updated imports and annotations in `UnitTestConfiguration` and `UnitCacheConfiguration` to reflect the migration of domain models and services (`Characteristics`, `Currency`, `Stage`, `GameMetadata`) to the `game.save` namespace. Added `GameMetadataService` mock for testing purposes.
