+++ lsadf_core/src/test/java/com/lsadf/core/unit/config/UnitTestConfiguration.java
@@ -15,11 +15,12 @@
  */
 package com.lsadf.core.unit.config;
 
-import com.lsadf.core.application.game.characteristics.CharacteristicsService;
-import com.lsadf.core.application.game.currency.CurrencyService;
-import com.lsadf.core.application.game.game_save.GameSaveService;
 import com.lsadf.core.application.game.inventory.InventoryService;
-import com.lsadf.core.application.game.stage.StageService;
+import com.lsadf.core.application.game.save.GameSaveService;
+import com.lsadf.core.application.game.save.characteristics.CharacteristicsService;
+import com.lsadf.core.application.game.save.currency.CurrencyService;
+import com.lsadf.core.application.game.save.metadata.GameMetadataService;
+import com.lsadf.core.application.game.save.stage.StageService;
 import com.lsadf.core.application.info.GlobalInfoService;
 import com.lsadf.core.application.search.SearchService;
 import com.lsadf.core.application.user.UserService;
@@ -57,6 +58,8 @@ public class UnitTestConfiguration {
   @MockBean private GlobalInfoService globalInfoService;
 
   @MockBean private CharacteristicsService characteristicsService;
+
+  @MockBean private GameMetadataService gameMetadataService;
 
   @MockBean private CurrencyService currencyService;
